General purpose delete photo + upload photo functions
